# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchCreated": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentRef"
            },
            "title": "Documents",
            "type": "array"
          },
          "events_url": {
            "description": "Server-sent events stream with live progress",
            "title": "Events Url",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "status",
          "documents",
          "events_url"
        ],
        "title": "BatchCreated",
        "type": "object"
      },
      "BatchDone": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "claims": {
            "title": "Claims",
            "type": "integer"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          },
          "type": {
            "const": "batch_done",
            "default": "batch_done",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "processed",
          "failed",
          "claims",
          "cost_usd"
        ],
        "title": "BatchDone",
        "type": "object"
      },
      "BatchFailed": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "type": {
            "const": "batch_failed",
            "default": "batch_failed",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "error"
        ],
        "title": "BatchFailed",
        "type": "object"
      },
      "BatchStarted": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "type": {
            "const": "batch_started",
            "default": "batch_started",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "total"
        ],
        "title": "BatchStarted",
        "type": "object"
      },
      "BatchView": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ClaimView"
            },
            "title": "Claims",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "documents": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DocumentView"
            },
            "title": "Documents",
            "type": "array"
          },
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "employee_id",
          "status",
          "total",
          "processed",
          "failed",
          "created_at"
        ],
        "title": "BatchView",
        "type": "object"
      },
      "Body_create_batch_v1_batches_post": {
        "properties": {
          "files": {
            "description": "Photos, PDFs or screenshots",
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_create_batch_v1_batches_post",
        "type": "object"
      },
...
      "ClaimMode": {
        "enum": [
          "trip",
          "period",
          "event",
          "allowance"
        ],
        "title": "ClaimMode",
        "type": "string"
      },
      "ClaimStatus": {
        "enum": [
          "draft",
          "needs_info",
          "ready",
          "submitted",
          "approved",
          "rejected"
        ],
        "title": "ClaimStatus",
        "type": "string"
      },
      "ClaimView": {
        "additionalProperties": false,
        "description": "A claim plus where it came from and how it will be routed.",
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "currency": {
            "default": "INR",
            "title": "Currency",
            "type": "string"
          },
          "document_ids": {
            "items": {
              "type": "string"
            },
            "title": "Document Ids",
            "type": "array"
          },
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "findings": {
            "description": "Claim-level findings",
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ClaimMode"
          },
          "open_questions": {
            "items": {
              "$ref": "#/components/schemas/OpenQuestion"
            },
            "title": "Open Questions",
            "type": "array"
          },
          "route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "status": {
            "$ref": "#/components/schemas/ClaimStatus",
            "default": "draft"
          },
          "submission_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Reference"
          },
          "title": {
            "description": "e.g. 'Pune trip 12-14 Aug' or 'Local conveyance Sept 2026'",
            "title": "Title",
            "type": "string"
          },
          "total": {
            "description": "Sum of document totals, in the claim currency",
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "id",
          "employee_id",
          "title",
          "mode",
          "document_ids",
          "total"
        ],
        "title": "ClaimView",
        "type": "object"
      },
      "ClaimsReady": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "type": {
            "const": "claims_ready",
            "default": "claims_ready",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "claim_ids"
        ],
        "title": "ClaimsReady",
        "type": "object"
      },
...
      "DocumentChecked": {
        "additionalProperties": false,
        "description": "Trust and policy checks finished for a document (in upload order).",
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "findings": {
            "title": "Findings",
            "type": "integer"
          },
          "trust_score": {
            "title": "Trust Score",
            "type": "integer"
          },
          "type": {
            "const": "document_checked",
            "default": "document_checked",
            "title": "Type",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "document_id",
          "trust_score",
          "verdict",
          "findings"
        ],
        "title": "DocumentChecked",
        "type": "object"
      },
      "DocumentExtracted": {
        "additionalProperties": false,
        "description": "A document has been read and categorised (the slow, parallel phase).",
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "category_confidence": {
            "title": "Category Confidence",
            "type": "number"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "type": {
            "const": "document_extracted",
            "default": "document_extracted",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "document_id",
          "filename",
          "position",
          "doc_type",
          "merchant",
          "total",
          "category",
          "category_confidence",
          "engine",
          "cached",
          "cost_usd"
        ],
        "title": "DocumentExtracted",
        "type": "object"
      },
      "DocumentFailed": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "type": {
            "const": "document_failed",
            "default": "document_failed",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "document_id",
          "filename",
          "error"
        ],
        "title": "DocumentFailed",
        "type": "object"
      },
      "DocumentRef": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename"
        ],
        "title": "DocumentRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Problem": {
        "description": "The body of every error response (``application/problem+json``).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI expense & reimbursement agent: from a pile of receipts to ready-to-submit.",
    "title": "ClaimPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/batches": {
      "post": {
        "operationId": "create_batch_v1_batches_post",
        "parameters": [
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_batch_v1_batches_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A file or the whole upload is too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Nothing to process, or an unsupported file"
          }
        },
        "summary": "Upload receipts and start processing",
        "tags": [
          "batches"
        ]
      }
    },
    "/v1/batches/{batch_id}": {
      "get": {
        "operationId": "get_batch_v1_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch status with its documents and claims",
        "tags": [
          "batches"
        ]
      }
    },
    "/v1/batches/{batch_id}/events": {
      "get": {
        "operationId": "batch_events_v1_batches__batch_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Server-sent events; each `data` is a PipelineEvent (see /history)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Live progress (server-sent events)",
        "tags": [
          "batches"
        ]
      }
    },
    "/v1/batches/{batch_id}/history": {
      "get": {
        "operationId": "batch_history_v1_batches__batch_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "description": "Skip this many events",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Skip this many events",
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "discriminator": {
                      "mapping": {
                        "batch_done": "#/components/schemas/BatchDone",
                        "batch_failed": "#/components/schemas/BatchFailed",
                        "batch_started": "#/components/schemas/BatchStarted",
                        "claims_ready": "#/components/schemas/ClaimsReady",
                        "document_checked": "#/components/schemas/DocumentChecked",
                        "document_extracted": "#/components/schemas/DocumentExtracted",
                        "document_failed": "#/components/schemas/DocumentFailed"
                      },
                      "propertyName": "type"
                    },
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/BatchStarted"
                      },
                      {
                        "$ref": "#/components/schemas/DocumentExtracted"
                      },
                      {
                        "$ref": "#/components/schemas/DocumentChecked"
                      },
                      {
                        "$ref": "#/components/schemas/DocumentFailed"
                      },
                      {
                        "$ref": "#/components/schemas/ClaimsReady"
                      },
                      {
                        "$ref": "#/components/schemas/BatchDone"
                      },
                      {
                        "$ref": "#/components/schemas/BatchFailed"
                      }
                    ]
                  },
                  "title": "Response Batch History V1 Batches  Batch Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Progress events so far (the same events as the SSE stream, as JSON)",
        "tags": [
          "batches"
        ]
      }
    },
...
  }
}